+++ src/preload/index.ts
@@ -102,7 +102,8 @@ const api = {
   getDataDir: () => ipcRenderer.invoke("app:get-data-dir"),
   getDefaultDataDir: () =>
     ipcRenderer.invoke("config:get-default-data-dir") as Promise<string>,
-  relaunch: () => ipcRenderer.invoke("app:relaunch"),
+  initDataDir: (dataDir: string) =>
+    ipcRenderer.invoke("app:init-data-dir", dataDir),
   getConfigDir: () =>
     ipcRenderer.invoke("app:get-config-dir") as Promise<string>,
   selectDirectory: () => ipcRenderer.invoke("app:select-directory"),